requirements.txt
flask
replicate
requests
pillow
opencv-python
tensorflow
numpy
scikit-learn
